pyproject.toml
[tool.poetry]
name = "resume-extractor"
version = "0.1.0"
description = "Resume extractor + qualitative LLM benchmarking"
authors = ["Sachin <sachin@example.com>"]
readme = "README.md"
packages = [{ include = "resume_extractor", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
pydantic = "^2.8.2"
pymupdf = "^1.27.1"
python-dotenv = "^1.2.1"
psutil = "^7.2.2"
pynvml = "^13.0.1"
openpyxl = "^3.1.5"
pandas = "^2.2.3"

langchain = "^0.3.0"
langchain-core = "^0.3.27"
langchain-community = "^0.3.0"
langchain-openai = "^0.2.14"
langchain-ollama = "^0.2.0"
openai = "^1.59.0"

tiktoken = "^0.12.0"
rapidfuzz = "^3.14.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[tool.poetry.scripts]
qualitative-test = "tests.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"